requirements.txt
Flask
Flask_Cors
requests
python-dotenv
jsonify